Fix GifPickerModal import path
- Corrected relative import path from '../utils/tenorApi' to '../../utils/tenorApi'
- Fixed Metro bundler resolution issue
- Ensured proper file structure navigation in reusable component
