+++ styles/HivePostScreenStyles.ts
@@ -176,73 +176,4 @@ export const HivePostScreenStyles = StyleSheet.create({
   commentsList: {
     marginTop: 8,
   },
-  gifModalContainer: {
-    flex: 1,
-    marginTop: 60,
-    borderTopLeftRadius: 20,
-    borderTopRightRadius: 20,
-  },
-  gifModalHeader: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-    borderBottomWidth: 1,
-  },
-  gifModalTitle: {
-    fontSize: 18,
-    fontWeight: '600',
-  },
-  gifSearchContainer: {
-    margin: 16,
-    borderBottomWidth: 1,
-  },
-  gifSearchInputContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderRadius: 25,
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-  },
-  gifSearchInput: {
-    flex: 1,
-    fontSize: 16,
-  },
-  gifClearButton: {
-    marginLeft: 8,
-  },
-  gifLoadingContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  gifEmptyContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 20,
-  },
-  gifEmptyText: {
-    fontSize: 16,
-    textAlign: 'center',
-    marginTop: 10,
-  },
-  gifGrid: {
-    padding: 16,
-  },
-  gifGridContent: {
-    justifyContent: 'space-between',
-  },
-  gifItem: {
-    flex: 1,
-    aspectRatio: 1,
-    margin: 4,
-    borderRadius: 8,
-    overflow: 'hidden',
-  },
-  gifImage: {
-    width: '100%',
-    height: '100%',
-  },
 });
